PR changes. Update location of fetchDocuments, reorg code for saving blob to computer
